refactor(engine): 单源化 decide 结果契约 schema
JsonSchema 类型迁至 @lorelum/shared（CLI 转发导出保持兼容）；engine 导出 decisionResultSchema 等契约 schema，CLI 引用同一对象，删除本地手写镜像（ADR 0008 §6 契约单源）。
